Refactor minctest macros and include order
- Renamed 'start' to 'lstart' in minctest macros for clarity.
- Replaced statement expressions with do-while loops in minctest macros.
- Corrected include order in test files for consistency.
- Changed 'exp' to 'expf' for float operations in test files.
- Added 'void' to main function signatures for better practice.

+++ test_frame_state.c
@@ -3,8 +3,8 @@
 #include "dna.h"
 #include "frame_state.h"
 #include "lprob.h"
-#include "span.h"
 #include "minctest.h"
+#include "span.h"
 
 static void run_frame_state1(float lprobs[], struct imm_span span)
 {
@@ -215,7 +215,7 @@ static void frame_state_lposterior(void)
         total = imm_lprob_add(total, lprob);
       }
     }
-    near((float)exp(total), 1.0);
+    near(expf(total), 1.0);
   }
 }
